* Replaced hard-coded project directory with /Users/david/TeamCitySlave/work/6a8478f881e6a6c1

+++ run_unity_tests.sh
@@ -1,13 +1,16 @@
 # Scripts to run the Unity Unit Tests on a CI (Jenkins) slave.
 
 # Removing the previous test results seems to be necessary to get the correct result for the next run.
-rm UnitTestResults.xml
+if [ -f UnitTestResults.xml ]
+then
+  rm UnitTestResults.xml
+fi
 
 # We need to force a refresh of the assets, otherwise the scripts will not be recompiled before the tests are run, and you will end up with the previous test code results
-/Applications/Unity/Unity.app/Contents/MacOS/Unity -batchmode -projectPath /Users/david/jenkins_slave/workspace/lwunity -executeMethod AssetDatabase.Refresh 
+/Applications/Unity/Unity.app/Contents/MacOS/Unity -batchmode -projectPath $PWD -executeMethod AssetDatabase.Refresh 
 
 # After refresh the current state of the code is reflected in the tests
-/Applications/Unity/Unity.app/Contents/MacOS/Unity -batchmode -projectPath /Users/david/jenkins_slave/workspace/lwunity -executeMethod UnityTest.UnitTestView.RunAllTestsBatch
+/Applications/Unity/Unity.app/Contents/MacOS/Unity -batchmode -projectPath $PWD -executeMethod UnityTest.UnitTestView.RunAllTestsBatch
 
 if [ ! -f UnitTestResults.xml ]
 then
